grpc-codes/Python3-Client/gnani_grpc_client/requirements.txt
grpcio
grpcio_tools
